separated network , model and views , created separate packages
