+++ .idea/modules.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project version="4">
+  <component name="ProjectModuleManager">
+    <modules>
+      <module fileurl="file://$PROJECT_DIR$/Android_App/app/src/androidTest/androidTest.iml" filepath="$PROJECT_DIR$/Android_App/app/src/androidTest/androidTest.iml" />
+      <module fileurl="file://$PROJECT_DIR$/Android_App/app/src/main/main.iml" filepath="$PROJECT_DIR$/Android_App/app/src/main/main.iml" />
+      <module fileurl="file://$PROJECT_DIR$/Android_App/app/src/test/test.iml" filepath="$PROJECT_DIR$/Android_App/app/src/test/test.iml" />
+    </modules>
+  </component>
+</project>
