+++ Android_App/app/src/main/java/com/infinitemakerz/heartdiseaseprediction/Network/ApiInterface.java
@@ -1,4 +1,6 @@
-package com.infinitemakerz.heartdiseaseprediction;
+package com.infinitemakerz.heartdiseaseprediction.Network;
+
+import com.infinitemakerz.heartdiseaseprediction.Model.Pojo;
 
 import java.util.List;
